requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyyaml==6.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
psutil==5.9.6
sshtunnel==0.4.0
paramiko==3.4.0
python-multipart==0.0.6
